not functional at the moment, splitting the files into multiple to make debugging and expanding easier
