+++ DictionaryGen/Details/Algorithm.md
