+++ DictionaryGen/Details/PreConfSyntax.fConfig
